ASM-2864 WWPN not added to the server
Manifest "createvol.pp" was using the first WWPN for creation of the server object. In case the server object already exists, this was never added to the server object.

This was not observed before as same WWPN values were used. In this case FCoE WWPN were used for creation of the server object and later FC WWPN were used for Server Object creation.

Manifest file is updated to pass all WWPN values during the mapping process. Existence of the WWPN values in the server object is handled by the puppet module