ASM-4419 Compellent volume creation fails due to incorrect storage profile name
Storage profile name "Low Priority" is not supported by SCv2000. Resulting the volume creation when the value of storage profile is not passed and default value is picked from the manifest file
Removed the default value. In absence of the value, volume is created by default with Storage profile High Priority