ASM-7971 EM discovery raising 400 for disconnected Compellent
The problem seems to be some of the Compellents in
EM were in “disconnected” state. The code that
tries to get the details of a Compellent from EM
was raising a 400 Bad Request error.

It probably makes sense for us to catch / rescue the
400 Bad Request error if it occurs during
discovery so that we don’t fail EM discovery.
However it means that our ASM EM inventory won’t
include any info on the Compellents that are in
disconnected state. Thus Compellent iSCSI
deployments to Compellents that are in this state
within EM will probably fail.